Geodata Functions

Geodata Functions

Usage,

add_map(data, maptype = "postcodes4", by = NULL, cut_northern_nl = TRUE, ...)

is.sf(sf_data)

cut_northern_nl(sf_data)

filter_sf(sf_data, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL)

filter_geolocation(sf_data, ..., col_zipcode = NULL)

latitude(sf_data)

longitude(sf_data)

Arguments

data

data to join left to the geodata

maptype

type of geometric data, must be one of: "gemeenten", "ggdregios", "inwoners_per_postcode_leeftijd", "jeugdregios", "nuts3regios", "postcodes", "postcodes4", "postcodes4_afstanden", "provincies", "veiligheidsregios"

sf_data

a data set of class 'sf'

xmin, xmax, ymin, ymax

coordination filters for sf_data

Value

An sf model. The column with geodata is always called "geometry".

Details

cut_northern_nl() cuts any geometry on the northern three provinces of the Netherlands.

filter_sf() filters an sf object on coordinates, and is internally used by cut_northern_nl().

latitude() and longitude() return these specific geographic properties of sf_data.

Examples

if (require("certeplot2")) {

  postcodes4 %>% 
    filter_geolocation(gemeente == "Tytsjerksteradiel") %>% 
    plot2()

}
#> Loading required package: certeplot2
#> Warning: there is no package called ‘certeplot2’