Aantal casussen binnen kilometerbereik bepalen

Met deze functie kan bepaald worden of er binnen een geografisch bereik van een postcode een minimaal aantal casussen (van bijv. een virus, isolaat of aandoening) voorgekomen is.

Usage,

cases_within_radius(...)

# S3 method for default
cases_within_radius(
  data,
  radius_km = 10,
  minimal_cases = 10,
  topic = NULL,
  slack = TRUE
)

# S3 method for sf
cases_within_radius(
  data_sf,
  radius_km = 10,
  minimal_cases = 10,
  column_zipcodes = "postcode",
  column_count = NULL,
  info = TRUE
)

# S3 method for cases_within_radius
format(x, ...)

# S3 method for cases_within_radius
format2(x, ...)

# S3 method for cases_within_radius
rbind(x, ...)

Arguments

data

Een data set met postcodes, gefilterd op positieve gevallen. Deze dataset moet een kolom 'postcode' bevatten met PC4-waarden. Zie Examples.

radius_km

Standaard is 10. De straal in kilometers waarbinnen het totaal minimal_cases voor moet komen, vanaf elke postcode. Dit is een thresholdparameter. Het kan voorkomen dat bijv. een postcode 10 km links van de postcode en een postcode 10 km rechts van postcode meegeteld worden. Dus de totale zoekdiameter rond elke postcode is 2x radius_km.

minimal_cases

Standaard is 10. Het minimaal aantal casussen dat voor moet komen binnen het geografisch gebied. Dit is een thresholdparameter.

topic

Standaard is leeg. Onderwerp om in geformatteerde tekst op te nemen. Dit kan de naam van een virus, bacterie of aandoening zijn.

slack

Resultaat naar Slack versturen, kanaal #early-warning.

data_sf

Een sf-object met in ieder geval deze kolommen: de postcode (instellen met column_zipcodes) en het aantal casussen (instellen met column_count).

column_zipcodes

Standaard is "postcode". De kolomnaam van de postcodes.

column_count

Standaard is eerste kolom die numeriek is en waarbij alle waarden binnen 0, 999 vallen.

info

Standaard is TRUE. Printen van geo-analyse en voortgang.

Value

Een list met de postcodes als namen, waarbij elk element een data.frame bevat met postcodes, aantallen en afstand in kilometers wanneer er van die postcode minimaal minimal_cases gevonden zijn binnen een straal van radius_km kilometer. Als dit niet het geval is, bevat het element NA. cases_within_radius() retourneert onzichtbaar de output van cases_within_radius.sf(). Gebruik format() of format2() om een printbaar formaat te maken, dat verzonden kan worden per mail of Slack. Met rbind() worden alle 'positieve' postcodes samengevoegd tot 1 data.frame, dat een geldig sf-object is.

Examples

if (FALSE) {
noro <- certedb_getmmb(where = db$t.testcode == "PRNORO",
                       zipcodes = TRUE)

# als je cases_within_radius() op 'normale' data gebruikt, worden de
# postcodes van de dataset 'postcodes4' genomen, wordt de inputdata
# gefilterd op positieven, zo nodig gegroepeerd op postcode en geteld,
# en wordt dan cases_within_radius.sf() gedraaid.
noro %>%
   cases_within_radius(radius_km = 5,
                       minimal_cases = 10,
                       topic = "norovirus",
                       slack = TRUE)

data.frame(ordernr = c("abc", "def"),
           postcode = c(8900, 8901),
           n = c(10, 20)) %>%
  add_map("pc4") %>%
  cases_within_radius(radius_km = 5, # <- dit is dus cases_within_radius.sf()
                      minimal_cases = 10) %>%
  format(topic = "SARS-CoV-2") %>%
  teams("early-warning")
}