Number of Inhabitants per Zip Code and Age

Number of Inhabitants per Zip Code and Age

inwoners_per_postcode_leeftijd

Format

A data.frame with 99,260 observations and 5 variables:

postcode
zip code, contains PC2, PC3 and PC4

leeftijd
age group per 5 years: 0-4, 5-9, ..., 90-94, 95+

inwoners
total number of inhabitants

inwoners_man
total number of male inhabitants

inwoners_vrouw
total number of female inhabitants

Source

The data in this data.frame are retrieved from, and publicly available at, Statistics Netherlands: StatLine, Centraal Bureau voor de Statistiek, 'Bevolking en leeftijd per postcode' (data set 83502NED), 1 januari 2021, https://opendata.cbs.nl.

Details

See the repository file to update this data set.

Examples

head(inwoners_per_postcode_leeftijd)
#> # A tibble: 6 × 5
#>   postcode leeftijd inwoners inwoners_man inwoners_vrouw
#>      <dbl> <ord>       <dbl>        <dbl>          <dbl>
#> 1       10 0-4         39925        20425          19500
#> 2       10 5-9         34790        17895          16895
#> 3       10 10-14       34095        17430          16665
#> 4       10 15-19       35950        17770          18180
#> 5       10 20-24       58850        26390          32460
#> 6       10 25-29       90430        42970          47455
str(inwoners_per_postcode_leeftijd)
#> tibble [99,260 × 5] (S3: tbl_df/tbl/data.frame)
#>  $ postcode      : num [1:99260] 10 10 10 10 10 10 10 10 10 10 ...
#>  $ leeftijd      : Ord.factor w/ 20 levels "0-4"<"5-9"<"10-14"<..: 1 2 3 4 5 6 7 8 9 10 ...
#>  $ inwoners      : num [1:99260] 39925 34790 34095 35950 58850 ...
#>  $ inwoners_man  : num [1:99260] 20425 17895 17430 17770 26390 ...
#>  $ inwoners_vrouw: num [1:99260] 19500 16895 16665 18180 32460 ...