Data Set with Dutch Zip Codes, Cities, Municipalities and Province

Data Set with Dutch Zip Codes, Cities, Municipalities and Province

postcodes

Format

A data.frame with 4,963 observations and 9 variables:

postcode
zip code, contains PC2, PC3 and PC4

inwoners
total number of inhabitants

inwoners_man
total number of male inhabitants

inwoners_vrouw
total number of female inhabitants

plaats
formal Dutch city name

gemeente
formal Dutch municipality name

provincie
formal Dutch province name

nuts3
Nomenclature of Territorial Units for Statistics, level 3 (in Dutch: COROP region, Coordinatie Commissie Regionaal OnderzoeksProgramma)

ggdregio
name of the regional GGD service (public healthcare service)

Source

The data in this data.frame are retrieved from, and publicly available at, Statistics Netherlands: StatLine, Centraal Bureau voor de Statistiek (CBS), 'Bevolking per geslacht per postcode' (data set 83503NED), 1 januari 2021, https://opendata.cbs.nl.

Details

See the repository file to update this data set.

Examples

head(postcodes)
#> # A tibble: 6 × 9
#>   postcode inwoners inwoners_man inwoners_vrouw plaats  gemeente provincie nuts3
#>      <dbl>    <dbl>        <dbl>          <dbl> <chr>   <chr>    <chr>     <chr>
#> 1       10   782445       388095         394350 Amster… Amsterd… Noord-Ho… Groo…
#> 2       11   306685       150760         155925 Amster… Amsterd… Noord-Ho… Groo…
#> 3       12   174380        85100          89285 Hilver… Hilvers… Noord-Ho… Het …
#> 4       13   259090       128150         130930 Almere  Almere   Flevoland Flev…
#> 5       14   230365       113120         117250 Bussum  Gooise … Noord-Ho… Het …
#> 6       15   181060        89675          91380 Zaandam Zaanstad Noord-Ho… Zaan…
#> # … with 1 more variable: ggdregio <chr>
str(postcodes)
#> tibble [4,963 × 9] (S3: tbl_df/tbl/data.frame)
#>  $ postcode      : num [1:4963] 10 11 12 13 14 15 16 17 18 19 ...
#>  $ inwoners      : num [1:4963] 782445 306685 174380 259090 230365 ...
#>  $ inwoners_man  : num [1:4963] 388095 150760 85100 128150 113120 ...
#>  $ inwoners_vrouw: num [1:4963] 394350 155925 89285 130930 117250 ...
#>  $ plaats        : chr [1:4963] "Amsterdam" "Amsterdam" "Hilversum" "Almere" ...
#>  $ gemeente      : chr [1:4963] "Amsterdam" "Amsterdam" "Hilversum" "Almere" ...
#>  $ provincie     : chr [1:4963] "Noord-Holland" "Noord-Holland" "Noord-Holland" "Flevoland" ...
#>  $ nuts3         : chr [1:4963] "Groot-Amsterdam" "Groot-Amsterdam" "Het Gooi en Vechtstreek" "Flevoland" ...
#>  $ ggdregio      : chr [1:4963] "GGD Amsterdam" "GGD Amsterdam" "GGD Gooi en Vechtstreek" "GGD Flevoland" ...