Distance from Zip Code to Zip Code

This data set was obtained by calculating the difference from the middle point of a zip code geometry to another zip code geometry (using the geo_postcodes4 data set and the sf package).

postcodes4_afstanden

Format

A data.frame with 562,330 observations and 3 variables:

postcode.x
zip code (PC4)

postcode.y
zip code (PC4)

afstand_km
distance in kilometres

Source

The data in this data.frame are retrieved from, and publicly available at, Statistics Netherlands:

Centraal Bureau voor de Statistiek (CBS), 'Gebiedsindelingen', GPKG 2022 v1, https://www.cbs.nl

Examples

head(postcodes4_afstanden)
#> # A tibble: 6 × 3
#>   postcode.x postcode.y afstand_km
#>   <chr>      <chr>           <dbl>
#> 1 7702       7702             0   
#> 2 7705       7702             3.04
#> 3 7715       7702             9.32
#> 4 7741       7702            14.4 
#> 5 7742       7702            17.2 
#> 6 7751       7702            18.4